requirements.txt
faiss-cpu
faiss-gpu-cu11
langchain
langchain-community
langchain-core
langchain-ollama
numpy
ollama
pypdf
python-dotenv
sentence-transformers
transformers